+++ app/components/primer/alpha/select_panel_element.ts
@@ -2,7 +2,7 @@ import {getAnchoredPosition} from '@primer/behaviors'
 import {controller, target} from '@github/catalyst'
 import {announceFromElement, announce} from '../aria_live'
 import {IncludeFragmentElement} from '@github/include-fragment-element'
-import type {PrimerTextFieldElement} from 'lib/primer/forms/primer_text_field'
+import type {PrimerTextFieldElement} from 'app/lib/primer/forms/primer_text_field'
 import type {AnchorAlignment, AnchorSide} from '@primer/behaviors'
 import '@oddbird/popover-polyfill'
 
@@ -165,7 +165,7 @@ export class SelectPanelElement extends HTMLElement {
   updateAnchorPosition() {
     // If the selectPanel is removed from the screen on resize close the dialog
     if (this && this.offsetParent === null) {
-      this.dialog.close()
+      this.hide()
     }
 
     if (this.invokerElement) {
@@ -314,11 +314,7 @@ export class SelectPanelElement extends HTMLElement {
         const itemContent = this.#getItemContent(item)
         if (!itemContent) continue
 
-        if (!this.isItemHidden(item) && !setZeroTabIndex) {
-          setZeroTabIndex = true
-        } else {
-          itemContent.setAttribute('tabindex', '-1')
-        }
+        itemContent.setAttribute('tabindex', '-1')
 
         // <li> elements should not themselves be tabbable
         item.removeAttribute('tabindex')
@@ -468,6 +464,12 @@ export class SelectPanelElement extends HTMLElement {
       // Remove data-ready so it can be set the next time the panel is opened
       this.dialog.removeAttribute('data-ready')
       this.invokerElement?.setAttribute('aria-expanded', 'false')
+      // When we close the dialog, clear the filter input
+      const fireSearchEvent = this.filterInputTextField.value.length > 0
+      this.filterInputTextField.value = ''
+      if (fireSearchEvent) {
+        this.filterInputTextField.dispatchEvent(new Event('input'))
+      }
 
       this.dispatchEvent(
         new CustomEvent('panelClosed', {
@@ -742,6 +744,8 @@ export class SelectPanelElement extends HTMLElement {
       return true
     }
 
+    if (!this.bannerErrorElement) return false
+
     return !this.bannerErrorElement.hasAttribute('hidden')
   }
 
@@ -882,6 +886,9 @@ export class SelectPanelElement extends HTMLElement {
     const itemContent = this.#getItemContent(item)
 
     if (this.selectVariant === 'single') {
+      // Don't check anything if we have an href
+      if (itemContent?.getAttribute('href')) return
+
       // disallow unchecking checked item in single-select mode
       if (!currentlyChecked) {
         for (const el of this.items) {