+++ app/lib/primer/forms/check_box_group.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Primer
+  module Forms
+    # :nodoc:
+    class CheckBoxGroup < BaseComponent
+      delegate :builder, :form, to: :@input
+
+      def initialize(input:)
+        @input = input
+        @input.add_label_classes("FormControl-label")
+        Primer::Forms::Utils.classify(@input.input_arguments)
+      end
+    end
+  end
+end