+++ test/components/alpha/select_panel_test.rb
@@ -18,7 +18,7 @@ module Primer
       end
 
       def test_does_not_render_a_filter_input
-        render_inline(Primer::Alpha::SelectPanel.new(show_filter: false))
+        render_inline(Primer::Alpha::SelectPanel.new(fetch_strategy: :local, show_filter: false))
 
         refute_selector("primer-text-field")
       end
@@ -95,11 +95,87 @@ module Primer
         assert_equal dialog_labelledby_id, header_id
       end
 
+      def test_custom_loading_label
+        render_preview(:custom_loading_label)
+
+        # Check that there's a loading label, but no description
+        assert_selector "svg[aria-label='Custom loading content... please wait...']"
+        refute_selector "svg[aria-label='Custom loading content... please wait...'][aria-describedby='select-panel-loading-description']"
+      end
+
+      def test_custom_loading_description
+        render_preview(:custom_loading_description)
+
+        # Check that there's a loading label and description
+        assert_selector "svg[aria-label='Custom loading content... please wait...'][aria-describedby='select-panel-loading-description']"
+        assert_selector "div", text: "This is a custom loading description", id: "select-panel-loading-description"
+      end
+
       def test_renders_close_button
         render_preview(:default)
 
         panel_id = page.find_css("select-panel").first.attributes["id"].value
         assert_selector "select-panel button[data-close-dialog-id='#{panel_id}-dialog']"
+      end
+
+      def test_raises_if_remote_strategy_and_hidden_filter_used_together
+        with_raise_on_invalid_options(true) do
+          error = assert_raises do
+            render_inline(Primer::Alpha::SelectPanel.new(fetch_strategy: :remote, show_filter: false))
+          end
+
+          assert_includes error.message, "Hiding the filter input with a remote fetch strategy is not permitted"
+        end
+      end
+
+      def test_raises_if_role_given
+        with_raise_on_invalid_options(true) do
+          error = assert_raises do
+            render_inline(Primer::Alpha::SelectPanel.new(role: :listbox))
+          end
+
+          assert_includes error.message, "Please avoid passing the `role:` argument"
+        end
+      end
+
+      def test_raises_if_role_given_to_item_slot
+        with_raise_on_invalid_options(true) do
+          error = assert_raises do
+            render_inline(Primer::Alpha::SelectPanel.new) do |panel|
+              panel.with_item(role: :option)
+            end
+          end
+
+          assert_includes error.message, "Please avoid passing the `role:` argument"
+        end
+      end
+
+      def test_does_not_raise_if_no_role_given_to_item_slot
+        render_inline(Primer::Alpha::SelectPanel.new) do |panel|
+          panel.with_item(label: "Foo")
+        end
+
+        assert_selector "select-panel"
+      end
+
+      def test_raises_if_role_given_to_avatar_item_slot
+        with_raise_on_invalid_options(true) do
+          error = assert_raises do
+            render_inline(Primer::Alpha::SelectPanel.new) do |panel|
+              panel.with_avatar_item(src: "camertron.jpg", username: "camertron", role: :option)
+            end
+          end
+
+          assert_includes error.message, "Please avoid passing the `role:` argument"
+        end
+      end
+
+      def test_does_not_raise_if_role_not_given_to_avatar_item_slot
+        render_inline(Primer::Alpha::SelectPanel.new) do |panel|
+          panel.with_avatar_item(src: "camertron.jpg", username: "camertron")
+        end
+
+        assert_selector(".avatar-small")
       end
     end
   end