+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # CHANGELOG
 
+## 0.34.0
+
+### Minor Changes
+
+- [#3062](https://github.com/primer/view_components/pull/3062) [`b45ef04`](https://github.com/primer/view_components/commit/b45ef044eb08788c7abff7cbd962da6d0ca56aa5) Thanks [@dombesz](https://github.com/dombesz)! - Trigger an input event when the primer-text-field clear button is pressed.
+
+### Patch Changes
+
+- [#3054](https://github.com/primer/view_components/pull/3054) [`4434ec0`](https://github.com/primer/view_components/commit/4434ec03af760a950060f58549243aa2842be045) Thanks [@camertron](https://github.com/camertron)! - [SelectPanel] Disallow passing `role:` argument
+
+- [#3077](https://github.com/primer/view_components/pull/3077) [`66488a3`](https://github.com/primer/view_components/commit/66488a30da6cccdf58367d0b20262af5afa597a1) Thanks [@camertron](https://github.com/camertron)! - [SelectPanel] Fix tab index issue in multi-select mode
+
+- [#3044](https://github.com/primer/view_components/pull/3044) [`b57b63a`](https://github.com/primer/view_components/commit/b57b63ab909a3afebd291f91d8e59db5ee9c0998) Thanks [@jamieshark](https://github.com/jamieshark)! - SelectPanel - design bug fixes. Adds a divider below the filter, adjusts min height for loading and error states, moves banner error above input filter."
+
+- [#3053](https://github.com/primer/view_components/pull/3053) [`871c818`](https://github.com/primer/view_components/commit/871c8184c8d3b70565052b419eec04ead2a894e0) Thanks [@camertron](https://github.com/camertron)! - [SelectPanel] Raise an error when remote fetch + hidden filter argument combo
+
+- [#3058](https://github.com/primer/view_components/pull/3058) [`269b02f`](https://github.com/primer/view_components/commit/269b02f221145af4acfb6fa8b4e15ef66c781b4c) Thanks [@langermank](https://github.com/langermank)! - `Label` padding updates to match PRC
+
+- [#3055](https://github.com/primer/view_components/pull/3055) [`ca4eae5`](https://github.com/primer/view_components/commit/ca4eae53c51abcd26b7fbaad78b6722e4c67a7e7) Thanks [@camertron](https://github.com/camertron)! - Remove OpenStruct from primer_octicon cop
+
+- [#3057](https://github.com/primer/view_components/pull/3057) [`96b42db`](https://github.com/primer/view_components/commit/96b42db2e0c0402008d2e00851e481d993b1506e) Thanks [@owenniblock](https://github.com/owenniblock)! - Don't select SelectPanel::Item when item is a link
+
+## 0.33.0
+
+### Minor Changes
+
+- [#3026](https://github.com/primer/view_components/pull/3026) [`5be0a0f`](https://github.com/primer/view_components/commit/5be0a0f1e60e02cb5140e159fd6b68dfd25194fa) Thanks [@camertron](https://github.com/camertron)! - Fire an event when banners are dismissed
+
+- [#3023](https://github.com/primer/view_components/pull/3023) [`3fb46ed`](https://github.com/primer/view_components/commit/3fb46ed1130f4a31e238b583de79db9d862c06a7) Thanks [@langermank](https://github.com/langermank)! - Bump `primer/primitives` v9
+
+- [#3022](https://github.com/primer/view_components/pull/3022) [`01cbbea`](https://github.com/primer/view_components/commit/01cbbea74f1b102cfffed69b83b464394c8019b9) Thanks [@camertron](https://github.com/camertron)! - Add a linter for migrating from SelectMenu -> SelectPanel
+
+### Patch Changes
+
+- [#3027](https://github.com/primer/view_components/pull/3027) [`5a205cd`](https://github.com/primer/view_components/commit/5a205cd63543c119b985e83e2800b8c2548cc051) Thanks [@camertron](https://github.com/camertron)! - Remove SelectPanel's backdrop
+
+- [#3021](https://github.com/primer/view_components/pull/3021) [`8a16336`](https://github.com/primer/view_components/commit/8a16336d254321f43969c9ed5b72861dc4c53c2e) Thanks [@camertron](https://github.com/camertron)! - When hiding a `Spinner`, also hide its screen reader text
+
+- [#3040](https://github.com/primer/view_components/pull/3040) [`612307d`](https://github.com/primer/view_components/commit/612307dd374352ccd71c8f82dbe7f065db54b9d4) Thanks [@camertron](https://github.com/camertron)! - Revert leading/trailing visuals for link component
+
+- [#3038](https://github.com/primer/view_components/pull/3038) [`6b4983e`](https://github.com/primer/view_components/commit/6b4983e1ecc29e7cf80cb68f0d274346882123b7) Thanks [@jamieshark](https://github.com/jamieshark)! - SelectPanel: check for filter input when loading remote items.
+
 ## 0.32.0
 
 ### Minor Changes